[neo]vim acts normally in tmux
This the first step in a move to transition away from complex multi-pane
windows managed in tmux, which I originally went with in order to have a
single window per project.

Now with tmux-sessionist and tmux-resurrect I am going to use a
session-per-project approach, with each project using one or more simple
1-2 pane windows. More complex splitting should probably be handled in
vim.